Fix debug statement warnings in cachetest.c

+++ tests/cachetest.c
@@ -117,20 +117,20 @@ int main(int argc, char *argv[])
 
         if (retcode != RIG_OK) { printf("Get vfo failed?? Err=%s\n", rigerror(retcode)); }
 
-        printf("%4dms: VFO = %s\n", elapsed_ms(&start, ELAPSED_GET), rig_strvfo(vfo));
+        printf("%4dms: VFO = %s\n", (int)elapsed_ms(&start, ELAPSED_GET), rig_strvfo(vfo));
 
         elapsed_ms(&start, ELAPSED_SET);
         retcode = rig_get_freq(my_rig, RIG_VFO_CURR, &freq);
 
         if (retcode != RIG_OK) { printf("Get freq failed?? Err=%s\n", rigerror(retcode)); }
 
-        printf("%4dms: VFO freq. = %.1f Hz\n", elapsed_ms(&start, ELAPSED_GET), freq);
+        printf("%4dms: VFO freq. = %.1f Hz\n", (int)elapsed_ms(&start, ELAPSED_GET), freq);
         elapsed_ms(&start, ELAPSED_SET);
         retcode = rig_get_mode(my_rig, RIG_VFO_CURR, &mode, &width);
 
         if (retcode != RIG_OK) { printf("Get mode failed?? Err=%s\n", rigerror(retcode)); }
 
-        printf("%4dms: Current mode = %s, width = %ld\n", elapsed_ms(&start,
+        printf("%4dms: Current mode = %s, width = %ld\n", (int)elapsed_ms(&start,
                 ELAPSED_GET),
                rig_strrmode(mode), width);
 
@@ -139,7 +139,7 @@ int main(int argc, char *argv[])
 
         if (retcode != RIG_OK) { printf("Get ptt failed?? Err=%s\n", rigerror(retcode)); }
 
-        printf("%4dms: ptt=%d\n", elapsed_ms(&start, ELAPSED_GET), ptt);
+        printf("%4dms: ptt=%d\n", (int)elapsed_ms(&start, ELAPSED_GET), ptt);
 
 #if 1
         elapsed_ms(&start, ELAPSED_SET);
@@ -147,12 +147,12 @@ int main(int argc, char *argv[])
 
         if (retcode != RIG_OK) { printf("Get split_vfo failed?? Err=%s\n", rigerror(retcode)); }
 
-        printf("%4dms: split=%d, tx_vfo=%s\n", elapsed_ms(&start, ELAPSED_GET), split,
+        printf("%4dms: split=%d, tx_vfo=%s\n", (int)elapsed_ms(&start, ELAPSED_GET), split,
                rig_strvfo(vfo));
 #endif
     }
 
-    printf("Elapsed %gsec\n", elapsed_ms(&startall, ELAPSED_GET) / 1000.0);
+    printf("Elapsed %gsec\n", (int)elapsed_ms(&startall, ELAPSED_GET) / 1000.0);
 
     rig_close(my_rig);
     return 0;
